+++ core/domain/interface/member/account.go
@@ -150,32 +150,32 @@ type (
 		GetPresentLog(id int32) *PresentLog
 
 		// 充值,客服操作时,需提供操作人(relateUser)
-		//ChargeForBalance(chargeType int32, title string, outerNo string, amount float32, relateUser int32) error
+		//ChargeForBalance(chargeType int32, title string, outerNo string, amount float32, relateUser int64) error
 
 		// 扣减余额
-		DiscountBalance(title string, outerNo string, amount float32, relateUser int32) error
+		DiscountBalance(title string, outerNo string, amount float32, relateUser int64) error
 
 		// 冻结余额
-		Freeze(title string, outerNo string, amount float32, relateUser int32) error
+		Freeze(title string, outerNo string, amount float32, relateUser int64) error
 
 		// 解冻金额
-		Unfreeze(title string, outerNo string, amount float32, relateUser int32) error
+		Unfreeze(title string, outerNo string, amount float32, relateUser int64) error
 
 		// 赠送金额,客服操作时,需提供操作人(relateUser)
-		//ChargeForPresent(title string, outerNo string, amount float32, relateUser int32) error
+		//ChargeForPresent(title string, outerNo string, amount float32, relateUser int64) error
 
 		// 赠送金额(指定业务类型)
-		//ChargePresentByKind(kind int32, title string, outerNo string, amount float32, relateUser int32) error
+		//ChargePresentByKind(kind int32, title string, outerNo string, amount float32, relateUser int64) error
 
 		// 扣减奖金,mustLargeZero是否必须大于0, 赠送金额存在扣为负数的情况
 		DiscountWallet(title string, outerNo string, amount float32,
-			relateUser int32, mustLargeZero bool) error
+			relateUser int64, mustLargeZero bool) error
 
 		// 冻结赠送金额
-		FreezeWallet(title string, outerNo string, amount float32, relateUser int32) error
+		FreezeWallet(title string, outerNo string, amount float32, relateUser int64) error
 
 		// 解冻赠送金额
-		UnfreezeWallet(title string, outerNo string, amount float32, relateUser int32) error
+		UnfreezeWallet(title string, outerNo string, amount float32, relateUser int64) error
 
 		// 流通账户余额变动，如扣除,amount传入负数金额
 		ChargeFlowBalance(title string, tradeNo string, amount float32) error
@@ -222,10 +222,10 @@ type (
 			amount float32, remark string) error
 
 		// 退款
-		Refund(accountKind int, kind int32, title string, outerNo string, amount float32, relateUser int32) error
+		Refund(accountKind int, kind int32, title string, outerNo string, amount float32, relateUser int64) error
 
 		// 充值
-		Charge(account int32, kind int32, title, outerNo string, amount float32, relateUser int32) error
+		Charge(account int32, kind int32, title, outerNo string, amount float32, relateUser int64) error
 
 		// 转账余额到其他账户
 		TransferBalance(kind int32, amount float32, tradeNo string, toTitle, fromTitle string) error
@@ -340,7 +340,7 @@ type (
 		// 手续费
 		CsnFee float32 `db:"csn_fee"`
 		// 关联操作人,仅在客服操作时,记录操作人
-		RelateUser int32 `db:"rel_user"`
+		RelateUser int64 `db:"rel_user"`
 		// 状态
 		State int32 `db:"state"`
 		// 备注
@@ -367,7 +367,7 @@ type (
 		// 手续费
 		CsnFee float32 `db:"csn_fee"`
 		// 关联操作人,仅在客服操作时,记录操作人
-		RelateUser int32 `db:"rel_user"`
+		RelateUser int64 `db:"rel_user"`
 		// 状态
 		State int32 `db:"state"`
 		// 备注
